moving 'startActivity' behaviour from ShareSessionBottomSheet.kt to SessionsController.kt
