+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveFragment.kt
@@ -49,7 +49,8 @@ class MobileActiveFragment : Fragment() {
             viewLifecycleOwner,
             settings,
             apiServiceFactory,
-            airbeamReconnector
+            airbeamReconnector,
+            context
         )
 
         if (sessionsRequested) {
