+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/dormant/MobileDormantFragment.kt
@@ -45,7 +45,8 @@ class MobileDormantFragment : Fragment() {
             viewLifecycleOwner,
             settings,
             apiServiceFactory,
-            childFragmentManager
+            childFragmentManager,
+            context
         )
 
         if (sessionsRequested) {
