+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/following/FollowingFragment.kt
@@ -45,7 +45,8 @@ class FollowingFragment : Fragment() {
             sessionsViewModel,
             viewLifecycleOwner,
             settings,
-            apiServiceFactory
+            apiServiceFactory,
+            context
         )
 
         if (sessionsRequested) {
